requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask_cors==5.0.0
flasgger==0.9.7.1
python-dotenv==0.21.0
openai==0.28.0
huggingface-hub==0.24.6
moviepy==1.0.3